Handle the case where toString() fails during error reporting.
In line 1973 we are reporting an error that happened while trying to initialize an object. Calling `toString()` on this object has a good chance of failing, and then the original error is not reported. I propose using the object's class name as a fallback when `toString()` fails.

The `== null` case is just for completeness.